Module weaver::graph::defs

Structs

CsvVector 🔒

Simple 3-dimensional vector struct for serializing points to csv.

Spinner

Weft

Weft is the main loop into which the warps are incorporated.

Constants

DISP_VECTORS

LAST_ROW

Traits

AbSumV

The sum of the absolute value of each scalar x, y and or z. Also known as the L1-Norm- manhattan distance from a to b where b is the origin (0, 0, 0) The L1 norm is calculated as the sum of the absolute vector values, where the absolute value of a scalar uses the notation |a1|. In effect, the norm is a calculation of the Manhattan distance from the origin of the vector space.

AbsOf

AddVec

AdjEdges

AlignToEdge

Bridge

Certify

Convert

CutYarn

DequeExtEnds

DrainFrom

ExtendBack

ExtendFront

ExtendThreads

GetEdges

GetUon

GetWeftWarps

InfoN

IsAdjacent

MakeEadjs

MakeEdges

Mirrored

OrientAscending

PinEnds

Prepare

ReverseIterator

ReverseIteratorSlice

ReversedVec

SerializeToCsv

Spin

Switch

Conveniece trait for an infinite iterator that doesn’t actually need unwrapping, so we’ll hide it.

Type Definitions

Adjacency

Count

Cushion

Edge

Edges

Loom

LoomSlice

LoomThread

Neighbors

Points

ScalarXyz

SignedIdx

Solution

Spool

Spun

Subtours

Tour

TourSlice

V

V2

VIMap

VecVert

Vert

Verts

Visited

Vs

Warp

Warps

Weights

Yarn

YarnEnds

Yarns

ZAdjacency

ZOrder

ZlevelVsMap