Trait weaver::graph::defs::AlignToEdge

pub trait AlignToEdge<T> {
    fn align_to(&mut self, edge: (V, V));
}

Required Methods

fn align_to(&mut self, edge: (V, V))

Align self to given edge such that the ends lhs of the edge and self match and the rhs of the edge and self match.